[components] Bug 1497808 - Create the PingMaker to collect data and assemble pings
This adds the PingMaker and slightly refactors the storage engines to
have implement the same interface, which is useful for handling the
ping generation. Additionally, this stubs out the "ping_info" section
that must be send with every ping. Even though a better solution would
have been for the fields of that section to dogfood our field types,
some of them are not planned for the MVP (e.g. DateTime), so I'm manually
handling this.