[components] Only execute module gradle task after 'clean' has finished running
On the new worker type, the old pattern of running 'clean' as part of the
module task is interacting poorly with parallel builds, and we're getting
IO errors. This change ensures we wait until clean is done before doing
any other work.

Locally this makes no difference (on Grisha's machine, at least) - and the
old pattern worked perfectly fine with the shared worker type.