[components] Closes https://github.com/mozilla-mobile/android-components/issues/1915: Memory leak caused by long-lived WebView instance
Co-authored-by: Sebastian Kaspari <s.kaspari@googlemail.com>
Co-authored-by: Christian Sadilek <christian.sadilek@gmail.com>

- The WebView instance is long-lived, as it's referenced in
SystemEngineSession, and we want to be able to keep engine
sessions around for the lifespan of the application.

- The WebView instance is created by the engine and using its
context (the application context).

- The EngineView, rendering engine sessions and WebViews,
is created using an activity context.

- The problem was that when the activity and EngineView were
destroyed, the WebView was not being detached and still
referenced the parent which referenced the activity context,
which would then not get gc'ed.

- We now detach the WebView when the EngineView is destroyed
and only destroy the WebView when the session is closed.

- This will allow re-using of engine sessions and their
WebViews in different activities and allow for proper
destruction.