[components] Added SQLCipher migration logic and updating the crypto handling
Copied the crypto code from autofill and adapted it for logins.  At some
point it would be nice to combine the common parts here, but I figure
that can be a later issue.

Added code to migrate from a SQLCipher database to the new plaintext
ones.  Choose the unimaginative "logins2.db" filename, feel free to pick
a different one.

Fixed some linter errors.