[components] Revert "Use the local server for testing"
This reverts commit cf03ecc73be1988e3aaa90edd353a458cdc2f9d9.

In Glean v50 `testGetValue` methods are actually not gated on test
mode (it's still not recommended though and absolutely IS still a
test-only API for all intents and purposes).

Further it turns out that calling `GleanTestLocalServer` in one test
might influence the other test and we're seeing intermittens where
`BaselinePingTest` is trying to upload to `localhost:0`, which of course
is invalid.
I have not yet figured out how we can properly separate out the state.
Alternatively we could merge the 2 tests into a single one.

Fixes https://github.com/mozilla-mobile/android-components/issues/11340