Bug 1804258 - Prevent overlapping of cancel requests for history reads
With more search providers querying the same history storage but only after
they try to cancel the in-progress queries it is possible for the cancel
requests to overlap and one search provider to cancel the queries of another.

To mitigate this a new method is available which will use the current input to
query the storage for as a way to drop new cancel requests for the same input.